Replaced the DeveloperTools.plist for a DeveloperToolsConstants class.
This has been replaced to the values previously stored in the plist can't be so easily inspected if someone decides they want to take a look into the app bundle. To use this, all that is needed is the DeveloperToolsConstants class should be extended in the main project, overriding the values needed.